refactor(bank-sync-tests): drop harness barrel, import from source files
Delete tests/bank-sync-tests/src/harness/index.ts (rule #30: no
re-export-only files). All 15 scenario files now import directly from
the specific harness source they need.
