+++ tests/bank-sync-tests/src/scenarios/consolidation/time-window-boundary.test.ts
@@ -2,7 +2,10 @@ import { describe, expect, it } from 'vitest';
 
 import { TRANSFER_PAIR_TIME_WINDOW_SECONDS, TransactionConsolidationTypeEnum } from '@budgie/contracts';
 
-import { fetchCanonicalsOfType, seedAccountPair, seedBankExpense, seedBankIncome, setupScenario } from '../../harness';
+import { fetchCanonicalsOfType } from '../../harness/fetch-canonicals';
+import { seedBankExpense, seedBankIncome } from '../../harness/seed-bank-pair';
+import { seedAccountPair } from '../../harness/seed-transfer-pair-fixture';
+import { setupScenario } from '../../harness/setup-scenario';
 
 import { transferConsolidationService } from '@app/sync/service/transfer-consolidation.service';
